@@ -4,26 +4,28 @@ plugins {
     id 'maven-publish'
     id 'kotlin-android'
     id 'kotlin-kapt'
-    id 'kotlin-android-extensions'
 }
 android {
+    namespace "com.amit.yoganet"
     compileSdkVersion 34
     buildToolsVersion "30.0.3"
 
 
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
+
+
 
 
     defaultConfig {
         applicationId "com.amit.yoganet"
         minSdkVersion 23
         targetSdkVersion 34
-        versionCode 21
-        versionName "1.7"
+        versionCode 22
+        versionName "1.8"
         multiDexEnabled true
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
@@ -38,19 +40,29 @@ android {
     buildFeatures {
         viewBinding = true
     }
+
+    viewBinding {
+        enabled = true
+    }
+
+    kotlinOptions {
+        jvmTarget = "17"
+    }
 }
+
+
 
 dependencies {
 
     //chat
 
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    api "androidx.constraintlayout:constraintlayout:2.1.3"
-    api "com.google.android.material:material:1.5.0"
+    api "androidx.constraintlayout:constraintlayout:2.1.4"
+    api "com.google.android.material:material:1.10.0"
     api "com.github.siyamed:android-shape-imageview:0.9.+@aar"
     api "de.hdodenhof:circleimageview:3.1.0"
     implementation "com.jakewharton:butterknife:10.2.3"
-    implementation 'androidx.coordinatorlayout:coordinatorlayout:1.1.0'
+    implementation 'androidx.coordinatorlayout:coordinatorlayout:1.2.0'
     annotationProcessor "com.jakewharton:butterknife-compiler:10.2.3"
     implementation "com.miguelcatalan:materialsearchview:1.4.0"
     implementation "com.jaredrummler:material-spinner:1.3.1"
@@ -67,10 +79,10 @@ dependencies {
 
     implementation 'com.android.support:multidex:1.0.3'
     implementation 'com.android.support.constraint:constraint-layout:2.0.4'
-    implementation 'com.google.android.gms:play-services-auth:20.1.0'
-    implementation 'com.google.android.material:material:1.4.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.2'
-    implementation "androidx.recyclerview:recyclerview:1.2.1"
+    implementation 'com.google.android.gms:play-services-auth:20.7.0'
+    implementation 'com.google.android.material:material:1.10.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation "androidx.recyclerview:recyclerview:1.3.2"
     implementation "androidx.cardview:cardview:1.0.0"
     implementation 'com.google.firebase:firebase-database:20.0.3'
     implementation 'com.google.firebase:firebase-firestore:24.0.0'
@@ -78,7 +90,7 @@ dependencies {
     implementation 'com.google.firebase:firebase-analytics:20.0.2'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.squareup.picasso:picasso:2.71828'
-   implementation  'androidx.appcompat:appcompat:1.4.0'
+   implementation  'androidx.appcompat:appcompat:1.6.1'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.2.0'
     //implementation 'com.google.firebase:firebase-ads:15.0.1'
 
@@ -99,6 +111,8 @@ dependencies {
     androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+
     implementation 'com.firebase:geofire-android:2.1.3'
     //noinspection GradleCompatible
     implementation 'com.android.support:appcompat-v7:25.4.0'
@@ -140,6 +154,9 @@ dependencies {
 
 }
 apply plugin: 'com.google.gms.google-services'
+apply plugin: 'com.android.application'
+
+
 repositories {
     mavenCentral()
 }
